Added custom env variables for email backend
This change also includes the changes to the Azure Pipelines
YAML file to build and push the Docker images using the Docker
Compose task instead of the Docker task used previously.

This change also introduces the docker-compose.outcome.yml file
which will help in handling the building of the custome images for
the cvat_server and cvat_ui services.